updated elastic search query to return number of words by language, changed number of words type from long to int

+++ src/main/java/org/dictionary/translator/StatTranslator.java
@@ -8,7 +8,7 @@ import org.dictionary.api.StatAPI;
 @Named
 public class StatTranslator {
 
-    public StatAPI toAPI(LanguageAPI language, long numWords) {
+    public StatAPI toAPI(LanguageAPI language, int numWords) {
         StatAPI stat = new StatAPI();
         stat.setLanguage(language);
         stat.setNumWords(numWords);
